Resolve merge conflicts with feat#9

+++ src/main/java/solvit/teachmon/domain/management/student/application/service/ManagementStudentService.java
@@ -3,26 +3,22 @@ package solvit.teachmon.domain.management.student.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import solvit.teachmon.domain.management.student.application.mapper.StudentMapper;
 import solvit.teachmon.domain.management.student.domain.entity.StudentEntity;
+import solvit.teachmon.domain.management.student.exception.StudentNotFoundException;
 import solvit.teachmon.domain.management.student.domain.repository.StudentRepository;
 import solvit.teachmon.domain.management.student.presentation.dto.request.StudentRequest;
-
-import java.time.LocalDate;
 
 @Service
 @RequiredArgsConstructor
 public class ManagementStudentService {
     private final StudentRepository studentRepository;
+    private final StudentMapper studentMapper;
 
     @Transactional
     public void createStudent(StudentRequest request) {
-        StudentEntity student = StudentEntity.builder()
-                .year(LocalDate.now().getYear())
-                .grade(request.grade())
-                .classNumber(request.classNumber())
-                .number(request.number())
-                .name(request.name())
-                .build();
+        StudentEntity student = studentMapper.toEntity(request);
+        student.setNowYear();
 
         studentRepository.save(student);
     }
@@ -30,13 +26,16 @@ public class ManagementStudentService {
     @Transactional
     public void updateStudent(Long studentId, StudentRequest request) {
         StudentEntity student = studentRepository.findById(studentId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 ID의 학생을 찾을 수 없습니다."));
+                .orElseThrow(StudentNotFoundException::new);
 
         student.changeInfo(request.grade(), request.classNumber(), request.number(), request.name());
     }
 
     @Transactional
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException();
+        }
         studentRepository.deleteById(studentId);
     }
 }
